Add PDF ticket export utility
