+++ package.json
@@ -13,27 +13,27 @@
   },
   "dependencies": {
     "@questlabs/react-sdk": "^2.2.4",
-    "@supabase/supabase-js": "^2.39.0",
     "@sentry/react": "^7.114.0",
-    "react": "^18.3.1",
-    "react-router-dom": "^7.1.0",
-    "react-dom": "^18.3.1",
-    "framer-motion": "^11.0.8",
-    "react-icons": "^5.4.0",
+    "@supabase/supabase-js": "^2.39.0",
+    "date-fns": "4.1.0",
     "echarts": "^5.5.0",
     "echarts-for-react": "^3.0.2",
-    "date-fns": "4.1.0",
+    "framer-motion": "^11.0.8",
     "html-to-image": "^1.11.11",
-    "jszip": "^3.10.1"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-icons": "^5.4.0",
+    "react-router-dom": "^7.1.0",
+    "pdf-lib": "^1.17.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.1",
     "@vitejs/plugin-react": "^4.3.1",
+    "autoprefixer": "^10.4.20",
     "eslint": "^9.9.1",
     "eslint-plugin-react-hooks": "^5.1.0-rc.0",
     "eslint-plugin-react-refresh": "^0.4.11",
     "globals": "^15.9.0",
-    "autoprefixer": "^10.4.20",
     "postcss": "^8.4.49",
     "tailwindcss": "^3.4.17",
     "vite": "^5.4.2"
